requirements.txt
absl-py==2.1.0
ale-py==0.8.1
astunparse==1.6.3
AutoROM==0.4.2
AutoROM.accept-rom-license==0.6.1
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools==5.3.3
certifi @ file:///croot/certifi_1720453481653/work/certifi
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.1.1
cycler==0.12.1
docker-pycreds==0.4.0
filelock @ file:///croot/filelock_1700591183607/work
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455532332/work
google-auth==2.32.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.64.1
gym==0.26.2
gym-notices==0.0.8
h5py==3.11.0
idna @ file:///croot/idna_1714398848350/work
importlib_metadata==8.0.0
importlib_resources==6.4.0
Jinja2 @ file:///croot/jinja2_1716993405101/work
joblib==1.4.2
keras==2.13.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.7.5
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
networkx @ file:///croot/networkx_1690561992265/work
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.1
pandas==2.0.3
pillow @ file:///croot/pillow_1714398848491/work
platformdirs==4.2.2
portalocker==2.10.1
protobuf==4.25.3
psutil==6.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.9.0.post0
pytz==2024.1
PyVirtualDisplay==3.0
PyYAML @ file:///croot/pyyaml_1698096049011/work
requests @ file:///croot/requests_1716902831423/work
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
sentry-sdk==2.9.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy @ file:///croot/sympy_1701397643339/work
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
threadpoolctl==3.5.0
torch==2.3.1
torchaudio==2.3.1
torchdata==0.7.1
torchtext==0.18.0
torchvision==0.18.1
tqdm==4.66.4
triton==2.3.1
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.1
urllib3 @ file:///croot/urllib3_1718912636303/work
wandb==0.17.4
Werkzeug==3.0.3
wrapt==1.16.0
xvfbwrapper @ file:///home/conda/feedstock_root/build_artifacts/xvfbwrapper_1648493254892/work
zipp==3.19.2
